MVCの採用
