+++ photo-werewolf-iOS/Auth/FirebaseAuth/HomeModel.swift
@@ -1,0 +1,13 @@
+//
+//  HomeModel.swift
+//  photo-werewolf-iOS
+//
+//  Created by 太田和希 on 2023/01/31.
+//
+
+import Foundation
+
+struct HomeModel {
+	var showingMakeRoomPopUp: Bool = false
+	var showingNameChangePopUp: Bool = false
+}
