+++ photo-werewolf-iOS/Features/StandardGame/ViewModels/WaitingRoomViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  WaitingRoomViewModel.swift
+//  photo-werewolf-iOS
+//
+//  Created by 太田和希 on 2023/02/02.
+//
+
+import Foundation
