+++ photo-werewolf-iOS/PhotoWerewolfIosApp.swift
@@ -24,7 +24,7 @@ struct 	PhotoWerewolfIosApp: App {
 
     var body: some Scene {
         WindowGroup {
-            HomeView()
+            HomeView(viewModel: HomeViewModel(model: HomeModel()))
         }
     }
 }
